requirements.txt
google-generativeai
pillow
protobuf
requests
python-dotenv